# api/openapi.yaml
openapi: 3.1.0
info:
  title: WAM — WhatsApp Marketing API
  version: 0.1.0
  description: |
    WAM — WhatsApp Marketing. Source of truth for HTTP APIs.

    * UI: `GET /api-docs/` (Swagger UI). Spec: `GET /api-docs/openapi.yaml`
      (alias `GET /openapi.yaml`). Legacy `/swagger*` 301s to `/api-docs/`.
    * Dashboard: `/dashboard /connect /contacts /templates /campaigns /analytics`
      (htmx, no React). Labels are embedded in Contacts (`/contacts#labels`), not a separate nav.
    * Connection pairing (QR/phone/logout/status), number check,
      one-off send, contacts/labels CRUD + import, templates CRUD + duplicate,
      campaigns + worker, per-campaign funnel, account overview, CSV export.
    * Persistence: SQLite (`WAM_DB_PATH`), shared with whatsmeow sqlstore.
      Auth: single admin (`WAM_ADMIN_PASSWORD`), session cookie.
    * SaaS path (Postgres `WAM_DATABASE_URL` + `WAM_APP_DATABASE_URL`):
      DB users + RBAC (`admin`/`user`, permission checks per handler),
      per-request org transactions (fail-closed RLS), per-account grants
      (`/api/v1/grants`, enforced in multi-account), signup/verify/reset,
      invites, multi-org sessions with switcher.
  contact:
    name: WAM
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://wam.example.com
    description: Production
...
paths:
...
  /api/v1/contacts:
    get:
      tags: [Contacts]
      operationId: listContacts
      summary: List contacts
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: group_id
          in: query
          schema: { type: string, format: uuid }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactPage'
    post:
      tags: [Contacts]
      operationId: createContact
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Phone already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/contacts/import:
    post:
      tags: [Contacts]
      operationId: importContacts
      summary: Bulk import CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '202':
          description: Import accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportReceipt'
...
  /api/v1/contacts/{id}:
    get:
      tags: [Contacts]
      operationId: getContact
      summary: Get contact
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '200':
          description: The contact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contacts]
      operationId: updateContact
      summary: Update contact
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Contacts]
      operationId: deleteContact
      summary: Delete contact
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '204': { description: Deleted }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/templates:
    get:
      tags: [Templates]
      operationId: listTemplates
      summary: List templates
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Template' }
    post:
      tags: [Templates]
      operationId: createTemplate
      summary: Create template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '409':
          description: Template name already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/templates/{id}:
    get:
      tags: [Templates]
      operationId: getTemplate
      summary: Get template
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '200':
          description: The template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Templates]
      operationId: updateTemplate
      summary: Update template
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Templates]
      operationId: deleteTemplate
      summary: Delete template
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '204': { description: Deleted }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/templates/{id}/duplicate:
    post:
      tags: [Templates]
      operationId: duplicateTemplate
      summary: Duplicate template
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    EntityId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PageCursor:
      name: cursor
      in: query
      schema: { type: string }
  schemas:
...
    Problem:
      title: RFC 9457 problem+json
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    Contact:
      type: object
      required: [id, phone, createdAt]
      properties:
        id: { type: string, format: uuid }
        phone: { type: string, description: E.164 phone number, example: '+15551234567' }
        name: { type: string, nullable: true }
        waOk: { type: boolean, nullable: true }
        groups:
          type: array
          items: { $ref: '#/components/schemas/Group' }
        createdAt: { type: string, format: date-time }
    ContactCreate:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: '+15551234567' }
        name: { type: string }
        group_ids:
          type: array
          items: { type: string, format: uuid }
    ContactPage:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Contact' }
        nextCursor: { type: string, nullable: true }
    ImportReceipt:
      type: object
      required: [accepted]
      properties:
        accepted: { type: integer }
        skipped: { type: integer }
    Group:
      type: object
      required: [id, name, contactCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string, nullable: true }
        contactCount: { type: integer }
...
    Template:
      type: object
      required: [id, name, body, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        category: { type: string, enum: [marketing, utility, authentication] }
        language: { type: string, example: en }
        header: { type: string, nullable: true }
        body: { type: string }
        footer: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    TemplateCreate:
      type: object
      required: [name, body]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        body: { type: string }
        category: { type: string, enum: [marketing, utility, authentication] }
        language: { type: string }
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...